+++ packages/firebase_performance/lib/firebase_performance.dart
@@ -1,0 +1,14 @@
+// Copyright 2017, the Flutter project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+library firebase_performance;
+
+import 'dart:async';
+import 'dart:collection';
+
+import 'package:flutter/services.dart';
+import 'package:flutter/foundation.dart';
+
+part 'src/firebase_performance.dart';
+part 'src/trace.dart';