+++ packages/google_maps_flutter/android/src/main/java/io/flutter/plugins/googlemaps/OnCameraMoveListener.java
@@ -7,7 +7,7 @@ package io.flutter.plugins.googlemaps;
 import com.google.android.gms.maps.model.CameraPosition;
 
 interface OnCameraMoveListener {
-  void onCameraMoveStarted(int reason);
+  void onCameraMoveStarted(boolean isGesture);
 
   void onCameraMove(CameraPosition newPosition);