+++ packages/google_maps_flutter/example/ios/Runner/AppDelegate.h
@@ -2,5 +2,4 @@
 #import <UIKit/UIKit.h>
 
 @interface AppDelegate : FlutterAppDelegate
-
 @end