+++ packages/google_maps_flutter/example/lib/map_ui.dart
@@ -7,22 +7,22 @@ import 'package:google_maps_flutter/google_maps_flutter.dart';
 
 import 'page.dart';
 
-const LatLngBounds sydneyBounds = const LatLngBounds(
+final LatLngBounds sydneyBounds = LatLngBounds(
   southwest: const LatLng(-34.022631, 150.620685),
   northeast: const LatLng(-33.571835, 151.325952),
 );
 
 class MapUiPage extends Page {
-  MapUiPage() : super(const Icon(Icons.map), "User interface");
+  MapUiPage() : super(const Icon(Icons.map), 'User interface');
 
   @override
   final GoogleMapOverlayController controller =
-      new GoogleMapOverlayController.fromSize(
+      GoogleMapOverlayController.fromSize(
     width: 300.0,
     height: 200.0,
-    options: const GoogleMapOptions(
+    options: GoogleMapOptions(
       cameraPosition: const CameraPosition(
-        target: const LatLng(-33.852, 151.211),
+        target: LatLng(-33.852, 151.211),
         zoom: 11.0,
       ),
       trackCameraPosition: true,
@@ -31,7 +31,7 @@ class MapUiPage extends Page {
 
   @override
   Widget build(BuildContext context) {
-    return new MapUiBody(controller);
+    return MapUiBody(controller);
   }
 }
 
@@ -41,12 +41,13 @@ class MapUiBody extends StatefulWidget {
   const MapUiBody(this.controller);
 
   @override
-  State<StatefulWidget> createState() => new MapUiBodyState();
+  State<StatefulWidget> createState() => MapUiBodyState();
 }
 
 class MapUiBodyState extends State<MapUiBody> {
   CameraPosition _position;
   GoogleMapOptions _options;
+  bool _isMoving;
 
   @override
   void initState() {
@@ -56,38 +57,38 @@ class MapUiBodyState extends State<MapUiBody> {
       setState(() {
         _options = mapController.options;
         _position = mapController.cameraPosition;
+        _isMoving = mapController.isCameraMoving;
       });
     });
     _options = mapController.options;
     _position = mapController.cameraPosition;
+    _isMoving = mapController.isCameraMoving;
   }
 
   Widget _compassToggler() {
-    return new FlatButton(
-      child:
-          new Text('${_options.compassEnabled ? 'disable' : 'enable'} compass'),
+    return FlatButton(
+      child: Text('${_options.compassEnabled ? 'disable' : 'enable'} compass'),
       onPressed: () {
         widget.controller.mapController.updateMapOptions(
-          new GoogleMapOptions(compassEnabled: !_options.compassEnabled),
+          GoogleMapOptions(compassEnabled: !_options.compassEnabled),
         );
       },
     );
   }
 
   Widget _latLngBoundsToggler() {
-    return new FlatButton(
-      child: new Text(
-        _options.latLngCameraTargetBounds.bounds == null
+    return FlatButton(
+      child: Text(
+        _options.cameraTargetBounds.bounds == null
             ? 'bound camera target'
             : 'release camera target',
       ),
       onPressed: () {
         widget.controller.mapController.updateMapOptions(
-          new GoogleMapOptions(
-            latLngCameraTargetBounds:
-                _options.latLngCameraTargetBounds.bounds == null
-                    ? const LatLngCameraTargetBounds(sydneyBounds)
-                    : LatLngCameraTargetBounds.unbounded,
+          GoogleMapOptions(
+            cameraTargetBounds: _options.cameraTargetBounds.bounds == null
+                ? CameraTargetBounds(sydneyBounds)
+                : CameraTargetBounds.unbounded,
           ),
         );
       },
@@ -95,13 +96,13 @@ class MapUiBodyState extends State<MapUiBody> {
   }
 
   Widget _zoomBoundsToggler() {
-    return new FlatButton(
-      child: new Text(_options.minMaxZoomPreference.minZoom == null
+    return FlatButton(
+      child: Text(_options.minMaxZoomPreference.minZoom == null
           ? 'bound zoom'
           : 'release zoom'),
       onPressed: () {
         widget.controller.mapController.updateMapOptions(
-          new GoogleMapOptions(
+          GoogleMapOptions(
             minMaxZoomPreference: _options.minMaxZoomPreference.minZoom == null
                 ? const MinMaxZoomPreference(12.0, 16.0)
                 : MinMaxZoomPreference.unbounded,
@@ -114,23 +115,23 @@ class MapUiBodyState extends State<MapUiBody> {
   Widget _mapTypeCycler() {
     final MapType nextType =
         MapType.values[(_options.mapType.index + 1) % MapType.values.length];
-    return new FlatButton(
-      child: new Text('change map type to $nextType'),
+    return FlatButton(
+      child: Text('change map type to $nextType'),
       onPressed: () {
         widget.controller.mapController.updateMapOptions(
-          new GoogleMapOptions(mapType: nextType),
+          GoogleMapOptions(mapType: nextType),
         );
       },
     );
   }
 
   Widget _rotateToggler() {
-    return new FlatButton(
-      child: new Text(
+    return FlatButton(
+      child: Text(
           '${_options.rotateGesturesEnabled ? 'disable' : 'enable'} rotate'),
       onPressed: () {
         widget.controller.mapController.updateMapOptions(
-          new GoogleMapOptions(
+          GoogleMapOptions(
             rotateGesturesEnabled: !_options.rotateGesturesEnabled,
           ),
         );
@@ -139,12 +140,12 @@ class MapUiBodyState extends State<MapUiBody> {
   }
 
   Widget _scrollToggler() {
-    return new FlatButton(
-      child: new Text(
+    return FlatButton(
+      child: Text(
           '${_options.scrollGesturesEnabled ? 'disable' : 'enable'} scroll'),
       onPressed: () {
         widget.controller.mapController.updateMapOptions(
-          new GoogleMapOptions(
+          GoogleMapOptions(
             scrollGesturesEnabled: !_options.scrollGesturesEnabled,
           ),
         );
@@ -153,12 +154,12 @@ class MapUiBodyState extends State<MapUiBody> {
   }
 
   Widget _tiltToggler() {
-    return new FlatButton(
-      child: new Text(
-          '${_options.tiltGesturesEnabled ? 'disable' : 'enable'} tilt'),
+    return FlatButton(
+      child:
+          Text('${_options.tiltGesturesEnabled ? 'disable' : 'enable'} tilt'),
       onPressed: () {
         widget.controller.mapController.updateMapOptions(
-          new GoogleMapOptions(
+          GoogleMapOptions(
             tiltGesturesEnabled: !_options.tiltGesturesEnabled,
           ),
         );
@@ -167,12 +168,12 @@ class MapUiBodyState extends State<MapUiBody> {
   }
 
   Widget _zoomToggler() {
-    return new FlatButton(
-      child: new Text(
-          '${_options.zoomGesturesEnabled ? 'disable' : 'enable'} zoom'),
+    return FlatButton(
+      child:
+          Text('${_options.zoomGesturesEnabled ? 'disable' : 'enable'} zoom'),
       onPressed: () {
         widget.controller.mapController.updateMapOptions(
-          new GoogleMapOptions(
+          GoogleMapOptions(
             zoomGesturesEnabled: !_options.zoomGesturesEnabled,
           ),
         );
@@ -182,24 +183,25 @@ class MapUiBodyState extends State<MapUiBody> {
 
   @override
   Widget build(BuildContext context) {
-    return new Column(
+    return Column(
       mainAxisAlignment: MainAxisAlignment.start,
       crossAxisAlignment: CrossAxisAlignment.stretch,
       children: <Widget>[
-        new Padding(
+        Padding(
           padding: const EdgeInsets.all(10.0),
-          child: new Center(
-            child: new GoogleMapOverlay(controller: widget.controller),
+          child: Center(
+            child: GoogleMapOverlay(controller: widget.controller),
           ),
         ),
-        new Column(
+        Column(
           children: <Widget>[
-            new Text('camera bearing: ${_position.bearing}'),
-            new Text(
+            Text('camera bearing: ${_position.bearing}'),
+            Text(
                 'camera target: ${_position.target.latitude.toStringAsFixed(4)},'
                 '${_position.target.longitude.toStringAsFixed(4)}'),
-            new Text('camera zoom: ${_position.zoom}'),
-            new Text('camera tilt: ${_position.tilt}'),
+            Text('camera zoom: ${_position.zoom}'),
+            Text('camera tilt: ${_position.tilt}'),
+            Text(_isMoving ? '(Camera moving)' : '(Camera idle)'),
             _compassToggler(),
             _latLngBoundsToggler(),
             _mapTypeCycler(),