+++ packages/firebase_storage/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.3.4
+
+* Added timeout properties to FirebaseStorage
+
+## 0.3.3
+
+* Added support for initialization with a custom Firebase app.
+
 ## 0.3.2
 
 * Added support for StorageReference `writeToFile`.